@@ -4,8 +4,8 @@
   "description": "Admin interface for creating Spark Daily editions",
   "private": true,
   "engines": {
-    "node": "6.10.1",
-    "npm": "3.10.5"
+    "node": "8.1.2",
+    "npm": "5.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -24,21 +24,22 @@
   },
   "homepage": "https://daily-composer.sparkthecause.com",
   "dependencies": {
-    "apollo-client": "^0.8.1",
+    "apollo-client": "^1.4.2",
     "daily-templates": "^0.8.0",
     "date-fns": "^1.28.5",
-    "graphql-tag": "^1.2.3",
+    "graphql-tag": "^2.4.0",
+    "prop-types": "^15.5.10",
     "react": "^15.4.2",
     "react-addons-update": "^15.4.2",
-    "react-apollo": "^0.9.0",
+    "react-apollo": "^1.4.2",
     "react-debounce-input": "^3.0.0",
     "react-dom": "^15.4.2",
     "react-router": "^2.8.1",
-    "react-sortable-hoc": "^0.4.5",
+    "react-sortable-hoc": "^0.6.3",
     "styled-components": "^2.0.0"
   },
   "devDependencies": {
-    "react-scripts": "0.8.5"
+    "react-scripts": "^1.0.7"
   },
   "eslintConfig": {
     "extends": "react-app"
