@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
 const colors = {
@@ -30,8 +31,8 @@ const Badge = ({ color, value }) =>
 ;
 
 Badge.propTypes = {
-  color: React.PropTypes.string.isRequired,
-  value: React.PropTypes.number
+  color: PropTypes.string.isRequired,
+  value: PropTypes.number
 };
 
 export default Badge;
